# swagger.yaml
openapi: 3.0.3
info:
  title: CompPare API
  description: API para gerenciamento de usuários, pastas, fotos e planos da plataforma CompPare
  version: 1.0.0
  contact:
    name: CompPare Support
    email: luispimenta.contato@gmail.com
servers:
  - url: https://api.comppare.com.br/api
    description: Servidor de produção
  - url: http://127.0.0.1:8000/api
    description: Servidor de desenvolvimento
paths:
...
  /pix/atualizar:
    post:
      summary: Atualiza status de cobranças PIX (webhook EFI)
      description: Endpoint utilizado pela EFI para notificar mudanças de status nas cobranças PIX recorrentes
      tags:
        - Pagamentos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recs
              properties:
                recs:
                  type: array
                  items:
                    type: object
                    required:
                      - idRec
                      - status
                    properties:
                      idRec:
                        type: string
                        example: "RR1026652320240821lab77511abf"
                        description: "ID único da cobrança recorrente"
                      status:
                        type: string
                        enum: [CRIADA, APROVADA, REJEITADA, CANCELADA]
                        example: "APROVADA"
                        description: "Novo status da cobrança"
                      atualizacao:
                        type: array
                        items:
                          type: object
                          properties:
                            status:
                              type: string
                            data:
                              type: string
                              format: date-time
                      ativacao:
                        type: object
                        properties:
                          tipoJornada:
                            type: string
                          dadosJornada:
                            type: object
                            properties:
                              txid:
                                type: string
      responses:
        '200':
          description: Atualização processada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  codRetorno:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Atualização de cobranças processada"
                  total_processados:
                    type: integer
                    example: 1
                  resultados:
                    type: array
                    items:
                      type: object
                      properties:
                        idRec:
                          type: string
                          example: "RR1026652320240821lab77511abf"
                        status:
                          type: string
                          example: "APROVADA"
                        status_anterior:
                          type: string
                          example: "ATIVA"
                        atualizado:
                          type: boolean
                          example: true
                        motivo:
                          type: string
                          example: "Pagamento não encontrado no sistema"
                          description: "Motivo quando atualizado=false"
        '400':
          description: Dados inválidos
        '500':
          description: Erro interno do servidor
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...